Yet anoth fix to compile on ia32
git-svn-id: https://devel.codership.com/svn/galera/trunk@295 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ wsdb/src/dump_trxfile.c
@@ -393,7 +393,7 @@ static int dump_block(struct block_info *bi) {
 	default:
             fprintf(stderr, "Bad rec type: %c\n", rec_type);
             fprintf(stderr, "at pos: %ld, block len: %d\n", 
-                    (bi->pos - bi->block), bi->block_len);
+                    (long)(bi->pos - bi->block), bi->block_len);
             exit(1);
             break;
 	}
